move member + template API contracts out of pages/api
