+++ front/lib/api/poke/templates.ts
@@ -1,7 +1,7 @@
 import { config } from "@app/lib/api/regions/config";
+import type { FetchAssistantTemplatesResponse } from "@app/lib/resources/template_resource";
 import { TemplateResource } from "@app/lib/resources/template_resource";
 import logger from "@app/logger/logger";
-import type { FetchAssistantTemplatesResponse } from "@app/pages/api/templates";
 import type { FetchAgentTemplateResponse } from "@app/pages/api/templates/[tId]";
 import type { ModelConfig } from "@app/types/assistant/models/types";
 import type {
