+++ front/lib/api/user.ts
@@ -14,6 +14,7 @@ import type { Result } from "@app/types/shared/result";
 import { Err, Ok } from "@app/types/shared/result";
 import type {
   LightWorkspaceType,
+  RoleType,
   UserType,
   UserTypeWithExtensionWorkspaces,
   UserTypeWithWorkspaces,
@@ -21,6 +22,26 @@ import type {
 
 import { MembershipResource } from "../resources/membership_resource";
 import { findWorkOSOrganizationsForUserId } from "./workos/organization_membership";
+
+export type GetMemberResponseBody = {
+  member: {
+    id: string;
+    username: string;
+    email: string;
+    firstName: string;
+    lastName: string | null;
+    fullName: string;
+    image: string | null;
+    revoked: boolean;
+    role: RoleType;
+    startAt: string | null;
+    endAt: string | null;
+  };
+};
+
+export type PostMemberResponseBody = {
+  member: UserTypeWithWorkspaces;
+};
 
 /**
  * Returns the acting user for an authenticated request. Falls back to looking up the user by
